This is a test image.
This is the same result that we obtained earlier using operator methods.
The function h(y) is thus a polynomial of order n, called a Hermite polynomial. We can determine the form of these polynomials either from the power-series solution or from our earlier result. The first three Hermite polynomials can be seen in the energy eigenfunctions. For n = 0, this equation says that a2/a0 = 0. Thus this series terminates after the first term and the first Hermite polynomial is simply a constant.
In this case we must set = 0, for otherwise the series starting with a j does not terminate and will behave as ey~ for large y. For n = 1, the situation is reversed. In this case, we must set a0 = 0 and the series starting with ax terminates after the first term. The second Hermite polynomial is simply y. Finally, for n = 2, we are back to the first case except that